updated AFNetworking to 1.1.0 and set main AFNetworking repo in podfile

+++ DXDAL/Code/Classes/HTTP/Data Providers/DXDALDataProviderMultipartForm.m
@@ -47,7 +47,7 @@
     __block NSString *videoURL = httpRequest.fileURLstring;
     __block NSInteger prevNotificationBytesCount = 0;
     
-    [operation setUploadProgressBlock:^(NSInteger bytesWritten, NSInteger totalBytesWritten, NSInteger totalBytesExpectedToWrite) {
+    [operation setUploadProgressBlock:^(NSUInteger bytesWritten, long long totalBytesWritten, long long totalBytesExpectedToWrite) {
         
         int part = totalBytesExpectedToWrite / 20;
